Fix properties of template members and add signature as astValue.
If a template class has a method then its definition also has to
be visited so we can jump on it and get its mangled name.

Moreover for function we add the signature as astValue. We use
astValue for textual representation of an AST node. This way we don't
need to get the properties of a function just to get its signature.