Adding C++ parser test
The first version of C++ parser test. This test checks what
nodes are inserted in CppAstNode table.

There is also some improvement in the C++ parser. The type locations
are added in the TypeLoc visitor functions. Their astType is not known
at this point, so two mapping data structures are used to gather this
information.

There was an extra unnecessary database entry for some function declarations
inserted with Read as astType.