Refactoring of ODB/SQL generation
The ODB generated files are placed under a model/include directory
under the current binary directory, except for the .cxx files. This
way the plugin creators don't have to use a special CMake function
for determining their location, because this structure will be common
for all models.
The plugin creators should'n be aware of the structure of the binary
directory.