Rewrite the request handling to work with new Mongoose
Mongoose 6.x line broke the API significantly, and multi-threading is
no longer officially supported, so we had to come up with a new
multithreading logic. Naive HTTP requests for static files is served
trivially by Mongoose itself, and requests to the Thrift API will use a
worker thread pool instead. This way the listener socket is not blocked
until a reply is created, and other users can still send requests.