CI: Switch to minimal image.
Only install required Boost components.
Extend Xenial and Bionic PostgreSQL build with postgresql server install.
Build and install CodeCompass in $HOME instead of $TRAVIS_BUILD_DIR, so the build and install directories won't be parsed on the self-parse test.
Run CodeCompass self-parse test on number of available CPU cores.
Replace the relative include path of Functions.cmake in webgui/InstallGUI.cmake with an absolute, so it works when the build directory is not in the source directory.