CallExpr can be applied to a function pointer
In an earlier commit (a8a1502ac68c1349281e34e2729cdc3988ccf6af) there was a
bugfix which related to fetching "is virtual" tag of a function pointer. Of
course function pointers can't have this flag. This previous commit fixed this
issue on server-side, however, it was the parser's fault that the symbol type
in a call expression was set to function even if it was a function pointer.